requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
boto3=1.12.40=pypi_0
botocore=1.15.40=pypi_0
ca-certificates=2020.4.5.1=hecc5488_0
certifi=2020.4.5.1=py36h9f0ad1d_0
cffi=1.14.0=py36h7a1dbc1_0
chardet=3.0.4=pypi_0
click=7.1.1=pypi_0
cpuonly=1.0=0
cycler=0.10.0=py36h009560c_0
dataclasses=0.7=pypi_0
docutils=0.15.2=pypi_0
filelock=3.0.12=pypi_0
freetype=2.9.1=ha9979f8_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.9=pypi_0
intel-openmp=2020.0=166
jieba=0.39=pyh9f0ad1d_1
jmespath=0.9.5=pypi_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.1.0=py36ha925a31_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.11.4=h1a1b453_0
libtiff=4.1.0=h56a325e_0
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36h64f37c6_0
mkl=2020.0=166
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.0.15=py36h14836fe_0
mkl_random=1.1.0=py36h675688f_0
ninja=1.9.0=py36h74a9793_0
numpy=1.18.1=py36h93ca92e_0
numpy-base=1.18.1=py36hc3f5095_1
olefile=0.46=py36_0
openssl=1.1.1f=hfa6e2cd_0
pillow=7.0.0=py36hcc1f983_0
pip=20.0.2=py36_1
protobuf=3.11.4=py36h97ec31f_1
pycparser=2.20=py_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py36h6538335_2
python=3.6.10=h9f7ef89_1
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.4.0=py3.6_cpu_0
pyyaml=5.3.1=py36he774522_0
qt=5.9.7=vc14h73c81de_0
regex=2020.4.4=pypi_0
requests=2.23.0=pypi_0
s3transfer=0.3.3=pypi_0
sacremoses=0.0.41=pypi_0
scikit-learn=0.22.1=py36h6288b17_0
scipy=1.4.1=py36h9439919_0
sentencepiece=0.1.85=pypi_0
setuptools=46.1.3=py36_0
sip=4.19.8=py36h6538335_0
six=1.14.0=py36_0
sqlite=3.31.1=he774522_0
tensorboardx=2.0=py_0
tk=8.6.8=hfa6e2cd_0
tokenizers=0.5.2=pypi_0
torchtext=0.5.0=pypi_0
torchvision=0.5.0=py36_cpu
tornado=6.0.4=py36he774522_1
tqdm=4.45.0=pypi_0
transformers=2.8.0=pypi_0
urllib3=1.25.9=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wheel=0.34.2=py36_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.5=h62dcd97_0
yaml=0.1.7=hc54c509_2
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
